Fix error not assignable to type 'ExportSpecifier'
Previously, running `tsc` from the repo root would give error:

toolchain/utils/node.ts(85,5): error TS2322: Type '{ type: "Literal";
    value: {}; raw: {}; __id: string; }' is not assignable to type
    'Node'.
  Type '{ type: "Literal"; value: {}; raw: {}; __id: string; }' is not
    assignable to type 'ExportSpecifier'.
  Property 'exported' is missing in type '{ type: "Literal"; value: {};
    raw: {}; __id: string; }'.

I'm not sure why, but it seems like ts expects the return value of this
function to be the 'ExportSpecifier' (defined by @types/estree). But,
this is clearly meant to return an object of type 'Identifier', or
'Literal'. Either way, all three classes are a subset of the class
es.Node, which is also the return type as per the function definition.

So, I looked into the commit history for this file back at the
source-academy2 repo, and found the syntax `as type`, and modified it to
fit our purposes (it used to be `as any`).

This is the last compilation error! Now to figure out how to run this
thing, and then to fix the runtime errors.