Add mutable variable declaration error
This is defined as `MutableVariableDeclarationError`, together with
the other errors in `toolchain/interpreter-errors.ts`. The reason for
choosing this location instead of the others (#10) was because there
were other errors that relate to rules that are later disabled for the
course e.g `VariableRedeclaration`.

Changes that were made to make this work:
- Adding of the error (that extends SourceError)
- modification of `defineVariable()` under `interpreter.ts` to take in
an additional parameter `kind` of the type `string | undefined`
- Modification of the calling of `defineVariable()` from variable
declaration and named function declaration. Note that the call at the function
declaration is the one that passes in `undefined`.

Results:
- A strict error (due to `ErrorSeverity.ERROR`) is given when a `let` or
`var` keyword is used to define variables. Error explanation and
elaboration (that can be generated from the class methods `explain()`
and `elaborate()`) are not specified, but this is basically issue #6