Make the repo an NPM package (#9)
* Change main file to index.ts

* Change main to pre-defined index

* Add index file at root

This is to allow 'slang' imports. The rootDir has been changed to '.'
since the index file exists here. In addition, the main file in
package.json has been changed to the root index file.

* Change index to src/index

This is because we will be exporting everything to the /dist directory
anyways.

* Add types export for tsconfig

Allows exporting types for the built files. Also changed rootDir

* Fix shadowing name error

When building, the Error type returned could also refer to Error defined
in node. So, I explicitly defined the return type.

* Change properties for publishing

- Change main and typings
- Add files to be included in publishing
- Add pre-publish script

* Use types property

Typings will be deprecated

* Change name and author

* Add organization name