Add curried function for display (#10)
* Add curried function for display

Must have accidentally changed it when incorporating externalBuiltIns

* Add alert to externalBuiltIns

Also curry prompt

* Add alert to CustomBuiltIns type

* Add list visualiser into externalBuiltins

Also fix a tyepdef for prompt and alert

* Change return type of alert to void

* Fix typescript errors

- Typo in 'value' ->> fixed to 'Value'
- non-exported List type casuing error

* Fix tsc error on private interface Suspended

* Map 'draw' to externalBuiltIns' visualiseList

The default value is to throw an error.

* Fix visualiseList return type

* Fix wrong function definitions showing up

This was becuase the __SOURCE__ property is read from the function that
is attached to the symbol. This means that the __SOURCE__ property must
be copied from the inner function to the outer one. I considered using
reflection (Object.* methods) but decided against it.