Fix tests not compiling
There was a problem with transpiling the files when testing, because of
the .js files used. I first added a transformer for the .js files (that
didnt have one at first, throwing an error at the `export` line in
parser.js). Then, I added the presets required, which were displayed as
missing presets when trying to test