Add Chapter 4 (#23)
* Add specified function definitions

Misc functions are to be imported after this

* Import type checking and error methods

Only some are used, but all of them are imported to keep this file up to
date with all the functions that were present before

* Remove duplicate error function

* Add metaCircularParser to Context

This Parser will be preserved for the lifetime of the Context, and this
ensures that it is reset together with the context as well.

* Add parse function

* Fix no typing error for misc

* Rename JSON.stringify -> stringify

* Remove unused Parser import

From rebasing away jison commits

* Add base parser file

* Format parser file

* Use .js for parser file

* Import and use parser for parse function

* Add parser jison source

* Add metacircular interpreter source

* Add doc to refer to source

* Fix tests not compiling

There was a problem with transpiling the files when testing, because of
the .js files used. I first added a transformer for the .js files (that
didnt have one at first, throwing an error at the `export` line in
parser.js). Then, I added the presets required, which were displayed as
missing presets when trying to test

* Update build to a script

This script will run the typescript build, and then copy over the
required .js file over. Note that to add more .js files, the build
script will need to be modified.

* Use && for failing tsc

* testing dev env

* Add .js files to interpreter folder

Oh my

* testing of martins dev setup for meta

* another try

* another try: getting really ugly

* getting uglier

* getting ready for prime time

* forgot removing tag

* trying to get => to work

* another attempt to fix arrow

* minor mods