Remove ArrowClosure which fixes another TCO bug
Arrow closures could not be TCO'd before as `apply` didn't handle it
correctly. Instead of making a bunch of changes to make `apply` work I
decided to simply make all `ArrowFunctionExpresion` evaluate to `Closure`
instead of `ArrowClosure`.

This causes some of the test cases to fail, but they should've failed
from the beginning as they were all demonstrating iterative processes.