Introduce mutability to language
Introduce mutability to language. Index table is now a 2D Array of name,
isVar pairs, where isVar determines whether a name is mutable or not. To
support mutability across functions, when an environment is extended, it
maintains a reference to the parent's environment to allow it to modify
names in that environment.