Update snapshots, see description for how to check:
I ran the following command to clean up the diff:

$ git show | grep -v '\(start\|end\|column\)' | grep '^[+-]' -C5 > cleaned-up.diff

This gets rid of most of the innocuous changes to the snapshot, where
the column number changed.

We do want to check that the line numbers are more or less correct,
which is why they aren't excluded from the grep search.

The result I got can be viewed here:

https://gist.github.com/thomastanck/b2950fdbd81b54fe29825c0a20e54c3e

I eyeballed this diff to see that the remaining changes are probably
just column number changes that couldn't be detected before.

There are a few places where the *ending* line number changed, but
that's fine because errors usually report the *starting* line number,
not the end.