Fix destructive renaming of variables when substituting within blocks
Fixed the destructive renaming of variables by modifying the variable declarator and function declaration
substituters, such that the id used in creation of the substituted variable declarator and function
declaration is a copy of the target.id, instead of the target id, which was used previously. Thus a new node
is created and used instead of the target.id would not be changed during renaming.
Also updated the test snapshots accordingly.

Resolves #872