Added function scanOutBoundedNames
The above function is used in functiondeclaration, functionexpression,
and arrowfunctionexpression substituters, to scan out all declared
names, as well as parameters in functions declared within the function
body. This is so that renaming will not clash with these inner
parameters which could cause capture. Added new testcases for these
as well.