Fix inconsistency of json imports by brute force (#1232)
There seems to be an inconsistency between how jest and how typescript
behaves when encountering imports of the form `import * as x from 'x.json'`
jest will set x = jsonobject,
but typescript will instead set x = { default: jsonobject }

This means that under typescript, we want `import x from 'x.json'`,
while under jest, we want `import * as x from 'x.json'`

This problem was hidden when transpiling to CommonJS modules before, which
behaves similarly to jest. But now that we are transpiling to es6,
typescript projects that depend on js-slang may now be exposed to this
inconsistency.

For now, we use brute force until the landscape changes or someone thinks of
a proper solution.