Modify return type of determineExecutionMethod to void
Currently, the determineExecutionMethod function does two things:
1) Sets the execution method on the context.
2) Returns whether the program is native runnable.

Upon further inspection, a program is native runnable if and only
if the execution method is set to "native". As a result, we should
defer to the check `context.executionMethod === "native"` to determine
whether or not a program is native runnable so that the context
acts as the single source of truth. This also has the added benefit
of making determineExecutionMethod have only a single responsibility -
determining the execution method as its name implies.