Refactor Source runner to operate on an AST instead of code (#1343)
* Fix typo 'hasDeuggerStatement' -> 'hasDebuggerStatement'

* Modify return type of determineExecutionMethod to void

Currently, the determineExecutionMethod function does two things:
1) Sets the execution method on the context.
2) Returns whether the program is native runnable.

Upon further inspection, a program is native runnable if and only
if the execution method is set to "native". As a result, we should
defer to the check `context.executionMethod === "native"` to determine
whether or not a program is native runnable so that the context
acts as the single source of truth. This also has the added benefit
of making determineExecutionMethod have only a single responsibility -
determining the execution method as its name implies.

* Refactor determineExecutionMethod to have a simpler control flow

* Update prelude handling comment to be more descriptive

* Move updating of context.unTypecheckedCode to sourceFilesRunner

* Update unTypedcheckedCode FIXME to be more descriptive

* Remove redundant pushing of code into unTypecheckedCode

The code in question should already have been pushed earlier on in
sourceFilesRunner & sourceRunner.

* Shift evaluation timeout increase check into sourceFilesRunner

This is so as to remove the runners' dependency on the code. It
also has the added benefit that it now considers all code files
and the entrypoint file path when determining whether or not a
particular run is a re-run that should have increased timeout.

* Store previous ASTs instead of code & fix test cases

* Fix testInContext not giving native runners a clean errors array

* Throw errors encountered while running instrumented code

* Exempt maximum call stack size exceeded errors from being thrown

* Make testForInfiniteLoop take in an AST instead of code

* Make fullJSRunner take in an AST instead of code

* Make sourceRunner take in an AST instead of code

* Throw error if unable to parse program in infinite loop tests