WIP: Configuring Env Viz to work with a step limit
* Added envSteps variable to context.runtime

* Passed options as an additional paramater to ECEvaluate and ECEMachine
  to access isPrelude option

* The evaluator now counts the number of steps (each command is one
  step)

The value in context.runtime.envSteps is the step limit the evaluator is
supposed to evaluate until.
If it is -1, the evaluator will evaluate to the end.

How the evaluator stops prematurely is by counting the number of steps
and inserts a dummy debugger statement to break the program when the
step limit is reached.